# sql: various allocation optimizations