--- a/client/util.go
+++ b/client/util.go
@@ -35,10 +35,12 @@ import (
 
 func marshalKey(k interface{}) (roachpb.Key, error) {
 	switch t := k.(type) {
-	case string:
-		return roachpb.Key(t), nil
+	case *roachpb.Key:
+		return *t, nil
 	case roachpb.Key:
 		return t, nil
+	case string:
+		return roachpb.Key(t), nil
 	case []byte:
 		return roachpb.Key(t), nil
 	}