* client: reduce allocations in Batch.initResult()
* sql: avoid allocations when passing keys to `client` methods
* sql: change marshalColumnValue() to return a roachpb.Value
* roachpb: avoid an allocation NewConditionalPut

```
name                  old time/op    new time/op    delta
KVInsert1_SQL-32         494µs ± 2%     494µs ± 3%     ~     (p=0.799 n=20+20)
KVInsert10_SQL-32        959µs ± 2%     947µs ± 2%   -1.17%  (p=0.001 n=19+20)
KVInsert100_SQL-32      5.22ms ± 3%    5.11ms ± 1%   -2.11%  (p=0.000 n=19+20)
KVInsert1000_SQL-32     52.2ms ± 4%    50.2ms ± 5%   -3.86%  (p=0.000 n=20+19)
KVInsert10000_SQL-32     669ms ± 8%     646ms ± 7%   -3.50%  (p=0.035 n=20+19)

name                  old alloc/op   new alloc/op   delta
KVInsert1_SQL-32        25.2kB ± 0%    25.1kB ± 0%   -0.32%  (p=0.000 n=20+20)
KVInsert10_SQL-32        101kB ± 0%     100kB ± 0%   -1.82%  (p=0.000 n=20+20)
KVInsert100_SQL-32       850kB ± 0%     832kB ± 0%   -2.17%  (p=0.000 n=18+20)
KVInsert1000_SQL-32     9.65MB ± 3%    9.51MB ± 4%   -1.37%  (p=0.001 n=19+20)
KVInsert10000_SQL-32     157MB ±29%     152MB ±18%     ~     (p=0.070 n=20+19)

name                  old allocs/op  new allocs/op  delta
KVInsert1_SQL-32           310 ± 0%       304 ± 0%   -1.94%  (p=0.000 n=17+20)
KVInsert10_SQL-32          825 ± 0%       754 ± 0%   -8.64%  (p=0.000 n=19+18)
KVInsert100_SQL-32       5.83k ± 0%     5.04k ± 0%  -13.48%  (p=0.000 n=20+20)
KVInsert1000_SQL-32      56.5k ± 1%     48.6k ± 2%  -13.87%  (p=0.000 n=19+20)
KVInsert10000_SQL-32      647k ±13%      563k ± 9%  -13.03%  (p=0.000 n=20+19)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/6378)
<!-- Reviewable:end -->