@dt 2016-04-28
Reviewed 1 of 1 files at r1, 5 of 5 files at r2, 2 of 2 files at r3, 1 of 1 files at r4.
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[client/batch.go, line 81 \[r4\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378#-KGSeBxcBK3zKXCYrwEz:-KGSeBxdCIr6T6dPzWva:1187583602) ([raw file](https://github.com/cockroachdb/cockroach/blob/1f9adca5226fa1dd4cd974c8a7d1997943b04748/client/batch.go#L81)):*
This is unreachable, right?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378)*
<!-- Sent from Reviewable.io -->