@RaduBerinde 2016-04-28
LGTM (modulo David's comment)

---

Reviewed 1 of 1 files at r1, 3 of 5 files at r2, 2 of 2 files at r3, 1 of 1 files at r4.
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378#-:-KGSmtBebtiuu0fWdlhp:-11731693)*
<!-- Sent from Reviewable.io -->