@petermattis 2016-04-28
Review status: all files reviewed at latest revision, 1 unresolved discussion, all commit checks successful.

---

*[client/batch.go, line 81 \[r4\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378#-KGSeBxcBK3zKXCYrwEz:-KGSnlHpRkXDcU3nO3zn:-420069405) ([raw file](https://github.com/cockroachdb/cockroach/blob/1f9adca5226fa1dd4cd974c8a7d1997943b04748/client/batch.go#L81)):*
Yes it is. Good catch.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378)*
<!-- Sent from Reviewable.io -->