@RaduBerinde 2016-04-28
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[client/batch.go, line 80 \[r8\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378#-KGSpPvzXk6nvS7hxAnd:-KGSpPvzXk6nvS7hxAne:610091905) ([raw file](https://github.com/cockroachdb/cockroach/blob/1e6fd044bdbadf3078190d9333286c9c023084b4/client/batch.go#L80)):*
this just became a very fancy way of writing `if numRows == 1`

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378)*
<!-- Sent from Reviewable.io -->