@petermattis 2016-04-28
Review status: all files reviewed at latest revision, 1 unresolved discussion, some commit checks pending.

---

*[client/batch.go, line 80 \[r8\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378#-KGSpPvzXk6nvS7hxAnd:-KGSq0JFXfD7tf-Mnlm3:1799698019) ([raw file](https://github.com/cockroachdb/cockroach/blob/1e6fd044bdbadf3078190d9333286c9c023084b4/client/batch.go#L80)):*
Yeah, I realized that too. The generated code is probably equivalent. I mildly prefer the switch format to the `if-else`, but happy to change this if you'd like.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/6378)*
<!-- Sent from Reviewable.io -->